+++ packages/braintree-integration/src/braintree-credit-card/braintree-credit-card-payment-strategy.ts
@@ -8,7 +8,6 @@ import {
 
 import {
     Address,
-    BraintreePaymentInstrument,
     isHostedInstrumentLike,
     isVaultedInstrument,
     MissingDataError,
@@ -18,7 +17,8 @@ import {
     OrderPaymentRequestBody,
     OrderRequestBody,
     PaymentArgumentInvalidError,
-    PaymentInitializeOptions, PaymentInstrument,
+    PaymentInitializeOptions,
+    PaymentInstrument,
     PaymentInstrumentMeta,
     PaymentIntegrationService,
     PaymentMethod,
@@ -171,7 +171,7 @@ export default class BraintreeCreditCardPaymentStrategy implements PaymentStrate
         payment: OrderPaymentRequestBody,
         billingAddress: Address,
         orderAmount: number,
-    ): Promise<BraintreePaymentInstrument & PaymentInstrumentMeta> {
+    ): Promise<PaymentInstrument & PaymentInstrumentMeta> {
         const { paymentData } = payment;
         const commonPaymentData = { deviceSessionId: this.deviceSessionId };