+++ packages/braintree-integration/src/braintree-request-sender.spec.ts
@@ -1,10 +1,10 @@
 import { createRequestSender, RequestSender } from '@bigcommerce/request-sender';
+
 import {
     ContentType,
     INTERNAL_USE_ONLY,
     SDK_VERSION_HEADERS,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
-
 import { getResponse } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
 
 import BraintreeRequestSender from './braintree-request-sender';