+++ packages/braintree-integration/src/mocks/braintree.mock.ts
@@ -8,7 +8,6 @@ import {
     BraintreeThreeDSecureOptions,
     PaypalButtonStyleColorOption,
 } from '@bigcommerce/checkout-sdk/braintree-utils';
-
 import {
     DefaultCheckoutButtonHeight,
     PaymentMethod,
@@ -35,7 +34,6 @@ export function getBraintreeAcceleratedCheckoutPaymentMethod(): PaymentMethod {
 
 export function getThreeDSecureMock(): BraintreeThreeDSecure {
     return {
-        // eslint-disable-next-line @typescript-eslint/no-unused-expressions
         verifyCard: (_options, callback: Braintree3DSVerifyCardCallback) => {
             if (callback) {
                 callback({ code: '' }, { nonce: 'fastlane_token_mock' });